requirements.txt
Flask==3.1.2
flask-cors==6.0.1
openai==1.30.1
gunicorn==21.2.0
python-dotenv==1.0.1
